% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadPlotServer}
\alias{downloadPlotServer}
\title{Register a Plot Download Module Server}
\usage{
downloadPlotServer(id, plot_reactive, filename_prefix = "plot")
}
\arguments{
\item{id}{Module id.}

\item{plot_reactive}{A reactive expression returning a plot object.}

\item{filename_prefix}{Prefix used for the downloaded file name.}
}
\value{
The result of \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.
}
\description{
Registers server logic for a simple plot download module.
}
