% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histonemod-app.R
\name{histonemod_app}
\alias{histonemod_app}
\title{Build the HistoneMod Shiny application object.}
\usage{
histonemod_app()
}
\value{
A `shiny.appobj` that can be launched with `shiny::runApp()`.
}
\description{
Build the HistoneMod Shiny application object.
}
