% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{histonemod_ui}
\alias{histonemod_ui}
\title{Build the HistoneMod User Interface}
\usage{
histonemod_ui()
}
\value{
A Shiny UI definition.
}
\description{
Creates the top-level Shiny UI for the HistoneMod application.
}
