% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoneMod-package.R
\docType{package}
\name{HistoneMod-package}
\alias{HistoneMod}
\alias{HistoneMod-package}
\title{HistoneMod package}
\description{
Provides a Shiny application and reusable helper functions for quantitative
analysis of histone post-translational modifications from peptide-level CSV
exports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jiehua1995/HistoneMod}
  \item Report bugs at \url{https://github.com/jiehua1995/HistoneMod/issues}
}

}
\author{
\strong{Maintainer}: Jie Hua \email{jie.Hua@bmc.med.lmu.de}

Authors:
\itemize{
  \item Jie Hua \email{jie.Hua@bmc.med.lmu.de}
}

Other contributors:
\itemize{
  \item Marco Borso [contributor]
  \item Beyza Bozdag [contributor]
  \item Axel Imhof [thesis advisor]
}

}
\keyword{internal}
