% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histonemod-app.R
\name{histonemod_app}
\alias{histonemod_app}
\title{Build the HistoneMod Shiny application}
\usage{
histonemod_app()
}
\value{
A \code{shiny.appobj} that can be launched with \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Constructs the Shiny application object used by HistoneMod.
}
