% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histonemod-app.R
\name{runHistoneMod}
\alias{runHistoneMod}
\title{Launch the HistoneMod Shiny application}
\usage{
runHistoneMod(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Invisibly returns the value from \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Launch the HistoneMod Shiny application
}
